Move common stuff to utils.py and add README
